extract what data are missing rf from phrma branch
